+++ plugins/remark-pdf/src/PdfViewer/PdfViewer.tsx
