+++ plugins/remark-pdf/src/remark-plugin/tests/assets/stairs.pdf
