+++ plugins/remark-pdf/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "noEmit": false
+    },
+    "rootDir": "${configDir}/src",
+    "outDir": "${configDir}/lib",
+    "include": ["src"],
+    "exclude": ["**/__tests__/**"]
+}
