+++ src/siteConfig/markdownPluginConfigs.ts
@@ -12,7 +12,7 @@ import mediaPlugin from '../plugins/remark-media/plugin';
 import detailsPlugin from '../plugins/remark-details/plugin';
 import pagePlugin from '../plugins/remark-page/plugin';
 import graphvizPlugin from '../plugins/remark-graphviz/plugin';
-import pdfPlugin from '../plugins/remark-pdf/plugin';
+import pdfPlugin from '../../plugins/remark-pdf/src/remark-plugin/index';
 import commentPlugin from '../plugins/remark-comments/plugin';
 import enumerateAnswersPlugin from '../plugins/remark-enumerate-components/plugin';
