+++ tsconfig.json
@@ -15,7 +15,7 @@
             "@tdev-stores/*": ["website/stores/*", "src/stores/*"],
             "@tdev-plugins/*": ["website/plugins/*", "src/plugins/*"],
             "@tdev-api/*": ["website/api/*", "src/api/*"],
-            "@tdev/*": ["website/*", "src/*"],
+            "@tdev/*": ["website/*", "src/*", "node_modules/@tdev/*"],
             "@tdev-original/*": ["src/*"],
             "@site/*": ["./*"]
         }
